Update join button
