+++ app/ui-message/client/messageBox.html
@@ -5,7 +5,7 @@
 				{{> messageBoxTyping rid=rid}}
 			{{/unless}}
 
-			{{#if canSend}}
+			{{#if isWritable}}
 				{{#if popupConfig}}
 					{{> messagePopupConfig popupConfig}}
 					{{> messagePopupSlashCommandPreview popupConfig}}
@@ -29,7 +29,7 @@
 							{{> icon block="rc-input__icon-svg" icon="send"}}
 						</span>
 					{{else}}
-						{{#if subscribed}}
+						{{#if canSend}}
 							{{> messageBoxAudioMessage rid=rid tmid=tmid}}
 							<span class="rc-message-box__action-menu js-action-menu" data-desktop>
 								{{#if actions}}
@@ -39,11 +39,12 @@
 								{{/if}}
 							</span>
 						{{else}}
-							<span class="rc-message-box__icon">
-								<strong><a href="#" class="js-join"> {{_"join"}}</a></strong>
-							</span>
+							<button class="js-join rc-button rc-button--primary rc-button--small rc-message-box__join-button">
+								{{_ "join"}}
+							</button>
 						{{/if}}
 					{{/unless}}
+
 					<span class="rc-message-box__action-label js-message-actions" data-small>
 						{{#each action in actions}}
 							<span class="js-message-action" data-id="{{action.id}}">
@@ -51,6 +52,7 @@
 							</span>
 						{{/each}}
 					</span>
+
 					<span class="rc-message-box__send js-message-action js-send" disabled="{{isSendIconDisabled}}" data-small>
 						{{> icon block="rc-message-box__action" icon="send"}}
 					</span>
